refactor(web): share the sidebar banner card with /m
The Star Agenta / changelog / trial cards lived in web/oss and rendered only on the
desktop rail. Move the model to @agenta/navigation and the renderer to
@agenta/navigation-ui (off antd: Typography -> span, Button -> EnhancedButton), then
fill the same footer slot from mobile's nav scope. EE keeps its subscription banners
through the unchanged additionalBannersAtom seam.
