+++ web/ee/package.json
@@ -24,6 +24,8 @@
         "@agenta/chat": "workspace:../packages/agenta-chat",
         "@agenta/entities": "workspace:../packages/agenta-entities",
         "@agenta/entity-ui": "workspace:../packages/agenta-entity-ui",
+        "@agenta/navigation": "workspace:../packages/agenta-navigation",
+        "@agenta/navigation-ui": "workspace:../packages/agenta-navigation-ui",
         "@agenta/observability-ui": "workspace:../packages/agenta-observability-ui",
         "@agenta/oss": "workspace:../oss",
         "@agenta/playground": "workspace:../packages/agenta-playground",
