+++ web/ee/src/components/SidebarBanners/state/atoms.ts
@@ -1,10 +1,10 @@
+import type {BannerConfig} from "@agenta/navigation"
 import {isBillingEnabled} from "@agenta/shared/api"
 import {dayjs} from "@agenta/shared/utils/dateTime"
 import duration from "dayjs/plugin/duration"
 import relativeTime from "dayjs/plugin/relativeTime"
 import {atom} from "jotai"
 
-import {BannerConfig} from "@/oss/components/SidebarBanners/types"
 import {isDemo} from "@/oss/lib/helpers/utils"
 import {freePlanSlugAtom} from "@/oss/state/access/atoms"
 import {urlAtom, URLState} from "@/oss/state/url"
