+++ web/oss/src/components/SidePanel/Subscription.tsx
@@ -1,8 +1,9 @@
+import {SidebarBanners} from "@agenta/navigation-ui"
+
 /**
- * OSS SidePanel Subscription component.
- * Re-exports the SidebarBanners component for sidebar integration.
+ * OSS SidePanel Subscription component — the shared banner card, unchanged.
  *
- * In EE mode, this is overridden by ee/src/components/SidePanel/Subscription.tsx
- * which injects subscription-based banners.
+ * In EE mode this module is shadowed by ee/src/components/SidePanel/Subscription.tsx,
+ * which injects subscription-based banners before rendering the same component.
  */
-export {default} from "@/oss/components/SidebarBanners"
+export default SidebarBanners
