+++ web/packages/agenta-navigation-ui/src/index.ts
@@ -23,3 +23,5 @@ export {
     type WorkflowPickerViewProps,
 } from "./WorkflowPickerView"
 export {SidebarToggleButton, type SidebarToggleButtonProps} from "./SidebarToggleButton"
+export {default as SidebarBanners} from "./SidebarBanners"
+export {default as SidebarBanner} from "./SidebarBanner"
