+++ web/packages/agenta-navigation/src/index.ts
@@ -9,6 +9,7 @@ export * from "./constants"
 export * from "./supportLinks"
 export * from "./state"
 export * from "./dynamic"
+export * from "./banners"
 
 export {useSidebarResize} from "./useSidebarResize"
